+++ test/components/agents/test_agent_breakpoints.py
@@ -15,7 +15,6 @@ from haystack.dataclasses import ChatMessage, ToolCall
 from haystack.dataclasses.breakpoints import AgentBreakpoint, AgentSnapshot, Breakpoint, ToolBreakpoint
 from haystack.tools import Tool
 from test.components.agents.test_agent import MockChatGenerator, weather_function
-from test.components.agents.test_agent_breakpoints_utils import agent_sync, mock_agent_with_tool_calls_sync
 
 
 @pytest.fixture
@@ -31,6 +30,15 @@ def weather_tool():
 @pytest.fixture
 def mock_chat_generator():
     generator = MockChatGenerator()
+    mock_run = MagicMock()
+    mock_run.return_value = {
+        "replies": [
+            ChatMessage.from_assistant(
+                "I'll help you check the weather.",
+                tool_calls=[{"tool_name": "weather_tool", "tool_args": {"location": "Berlin"}}],
+            )
+        ]
+    }
     mock_run_async = AsyncMock()
     mock_run_async.return_value = {
         "replies": [
@@ -41,9 +49,13 @@ def mock_chat_generator():
         ]
     }
 
+    def mock_run_with_tools(messages, tools=None, **kwargs):
+        return mock_run.return_value
+
     async def mock_run_async_with_tools(messages, tools=None, **kwargs):
         return mock_run_async.return_value
 
+    generator.run = mock_run_with_tools
     generator.run_async = mock_run_async_with_tools
     return generator
 
@@ -64,8 +76,7 @@ def mock_agent_with_tool_calls(monkeypatch, weather_tool):
     generator = MockChatGenerator()
     mock_messages = [
         ChatMessage.from_assistant("First response"),
-        ChatMessage.from_assistant(
-            tool_calls=[ToolCall(tool_name="weather_tool", arguments={"location": "Berlin"})]),
+        ChatMessage.from_assistant(tool_calls=[ToolCall(tool_name="weather_tool", arguments={"location": "Berlin"})]),
     ]
     agent = Agent(chat_generator=generator, tools=[weather_tool], max_agent_steps=10)  # Increase max steps
     agent.warm_up()
@@ -80,21 +91,21 @@ def debug_path(tmp_path):
 
 
 class TestAgentBreakpoints:
-    def test_run_with_chat_generator_breakpoint(self, agent_sync):
+    def test_run_with_chat_generator_breakpoint(self, agent):
         messages = [ChatMessage.from_user("What's the weather in Berlin?")]
         chat_generator_bp = Breakpoint(component_name="chat_generator", visit_count=0)
         agent_breakpoint = AgentBreakpoint(break_point=chat_generator_bp, agent_name="test_agent")
         with pytest.raises(BreakpointException) as exc_info:
-            agent_sync.run(messages=messages, break_point=agent_breakpoint)
+            agent.run(messages=messages, break_point=agent_breakpoint)
         assert exc_info.value.component == "chat_generator"
         assert "messages" in exc_info.value.inputs["chat_generator"]["serialized_data"]
 
-    def test_run_with_tool_invoker_breakpoint(self, mock_agent_with_tool_calls_sync):
+    def test_run_with_tool_invoker_breakpoint(self, mock_agent_with_tool_calls):
         messages = [ChatMessage.from_user("What's the weather in Berlin?")]
         tool_bp = ToolBreakpoint(component_name="tool_invoker", visit_count=0, tool_name="weather_tool")
         agent_breakpoint = AgentBreakpoint(break_point=tool_bp, agent_name="test_agent")
         with pytest.raises(BreakpointException) as exc_info:
-            mock_agent_with_tool_calls_sync.run(messages=messages, break_point=agent_breakpoint)
+            mock_agent_with_tool_calls.run(messages=messages, break_point=agent_breakpoint)
 
         assert exc_info.value.component == "tool_invoker"
         assert {"chat_generator", "tool_invoker"} == set(exc_info.value.inputs.keys())
@@ -102,14 +113,14 @@ class TestAgentBreakpoints:
         assert "serialized_data" in exc_info.value.inputs["chat_generator"]
         assert "messages" in exc_info.value.inputs["chat_generator"]["serialized_data"]
 
-    def test_resume_from_chat_generator(self, agent_sync, tmp_path):
+    def test_resume_from_chat_generator(self, agent, tmp_path):
         messages = [ChatMessage.from_user("What's the weather in Berlin?")]
         debug_path = str(tmp_path / "debug_snapshots")
         chat_generator_bp = Breakpoint(component_name="chat_generator", visit_count=0, snapshot_file_path=debug_path)
         agent_breakpoint = AgentBreakpoint(break_point=chat_generator_bp, agent_name="test_agent")
 
         try:
-            agent_sync.run(messages=messages, break_point=agent_breakpoint)
+            agent.run(messages=messages, break_point=agent_breakpoint)
         except BreakpointException:
             pass
 
@@ -117,7 +128,7 @@ class TestAgentBreakpoints:
         assert len(snapshot_files) > 0
         latest_snapshot_file = str(max(snapshot_files, key=os.path.getctime))
 
-        result = agent_sync.run(
+        result = agent.run(
             messages=[ChatMessage.from_user("Continue from where we left off.")],
             snapshot=load_pipeline_snapshot(latest_snapshot_file).agent_snapshot,
         )
@@ -126,7 +137,7 @@ class TestAgentBreakpoints:
         assert "last_message" in result
         assert len(result["messages"]) > 0
 
-    def test_resume_from_tool_invoker(self, mock_agent_with_tool_calls_sync, tmp_path):
+    def test_resume_from_tool_invoker(self, mock_agent_with_tool_calls, tmp_path):
         messages = [ChatMessage.from_user("What's the weather in Berlin?")]
         debug_path = str(tmp_path / "debug_snapshots")
         tool_bp = ToolBreakpoint(
@@ -135,9 +146,7 @@ class TestAgentBreakpoints:
         agent_breakpoint = AgentBreakpoint(break_point=tool_bp, agent_name="test_agent")
 
         try:
-            mock_agent_with_tool_calls_sync.run(
-                messages=messages, break_point=agent_breakpoint, agent_name="test_agent"
-            )
+            mock_agent_with_tool_calls.run(messages=messages, break_point=agent_breakpoint, agent_name="test_agent")
         except BreakpointException:
             pass
 
@@ -145,7 +154,7 @@ class TestAgentBreakpoints:
         assert len(snapshot_files) > 0
         latest_snapshot_file = str(max(snapshot_files, key=os.path.getctime))
 
-        result = mock_agent_with_tool_calls_sync.run(
+        result = mock_agent_with_tool_calls.run(
             messages=[ChatMessage.from_user("Continue from where we left off.")],
             snapshot=load_pipeline_snapshot(latest_snapshot_file).agent_snapshot,
         )
@@ -154,25 +163,27 @@ class TestAgentBreakpoints:
         assert "last_message" in result
         assert len(result["messages"]) > 0
 
-    def test_invalid_combination_breakpoint_and_pipeline_snapshot(self, mock_agent_with_tool_calls_sync):
+    def test_invalid_combination_breakpoint_and_pipeline_snapshot(self, mock_agent_with_tool_calls):
         messages = [ChatMessage.from_user("What's the weather in Berlin?")]
         tool_bp = ToolBreakpoint(component_name="tool_invoker", visit_count=0, tool_name="weather_tool")
         agent_breakpoint = AgentBreakpoint(break_point=tool_bp, agent_name="test_agent")
         with pytest.raises(ValueError, match="break_point and snapshot cannot be provided at the same time"):
-            mock_agent_with_tool_calls_sync.run(
-                messages=messages, break_point=agent_breakpoint, snapshot={"some": "snapshot"}
+            mock_agent_with_tool_calls.run(
+                messages=messages,
+                break_point=agent_breakpoint,
+                snapshot=AgentSnapshot(component_inputs={}, component_visits={}, break_point=agent_breakpoint),
             )
 
-    def test_breakpoint_with_invalid_component(self, mock_agent_with_tool_calls_sync):
+    def test_breakpoint_with_invalid_component(self, mock_agent_with_tool_calls):
         invalid_bp = Breakpoint(component_name="invalid_breakpoint", visit_count=0)
         with pytest.raises(ValueError):
             AgentBreakpoint(break_point=invalid_bp, agent_name="test_agent")
 
-    def test_breakpoint_with_invalid_tool_name(self, mock_agent_with_tool_calls_sync):
+    def test_breakpoint_with_invalid_tool_name(self, mock_agent_with_tool_calls):
         tool_breakpoint = ToolBreakpoint(component_name="tool_invoker", visit_count=0, tool_name="invalid_tool")
         with pytest.raises(ValueError, match="Tool 'invalid_tool' is not available in the agent's tools"):
             agent_breakpoints = AgentBreakpoint(break_point=tool_breakpoint, agent_name="test_agent")
-            mock_agent_with_tool_calls_sync.run(
+            mock_agent_with_tool_calls.run(
                 messages=[ChatMessage.from_user("What's the weather in Berlin?")], break_point=agent_breakpoints
             )
 
@@ -258,9 +269,9 @@ class TestAsyncAgentBreakpoints:
         agent_breakpoint = AgentBreakpoint(break_point=tool_bp, agent_name="test")
         with pytest.raises(ValueError, match="break_point and snapshot cannot be provided at the same time"):
             await mock_agent_with_tool_calls.run_async(
-                messages=messages, break_point=agent_breakpoint, snapshot=AgentSnapshot(
-                    component_inputs={}, component_visits={}, break_point=agent_breakpoint
-                )
+                messages=messages,
+                break_point=agent_breakpoint,
+                snapshot=AgentSnapshot(component_inputs={}, component_visits={}, break_point=agent_breakpoint),
             )
 
     @pytest.mark.asyncio
